+++ Sources/RichTextKit/RichTextViewComponent.swift
@@ -39,51 +39,31 @@ public protocol RichTextViewComponent: AnyObject,
     RichTextStyleWriter
 {
         
-    /**
-     The text view's frame.
-     */
+    /// The text view's frame.
     var frame: CGRect { get }
 
-    /**
-     The style to use when highlighting text in the view.
-     */
+    /// The style to use when highlighting text in the view.
     var highlightingStyle: RichTextHighlightingStyle { get set }
 
-    /**
-     The image configuration used by the rich text view.
-     */
+    /// The image configuration used by the rich text view.
     var imageConfiguration: RichTextImageConfiguration { get set }
 
-    /**
-     Whether or not the text view is the first responder.
-     */
+    /// Whether or not the text view is the first responder.
     var isFirstResponder: Bool { get }
 
-    /**
-     The text view's mutable attributed string, if any.
-     */
+    /// The text view's mutable attributed string, if any.
     var mutableAttributedString: NSMutableAttributedString? { get }
 
-    /**
-     The spacing between the text view's edge and its text.
-     */
+    /// The spacing between the text view's edge and its text.
     var textContentInset: CGSize { get set }
 
-    /**
-     The text view current typing attributes.
-     */
+    /// The text view current typing attributes.
     var typingAttributes: RichTextAttributes { get set }
 
 
     // MARK: - Setup
 
-    /**
-     Setup the view with text and a ``RichTextDataFormat``.
-
-     - Parameters:
-       - text: The text to edit with the text view.
-       - format: The rich text format to use.
-     */
+    /// Setup the view with a text and ``RichTextDataFormat``.
     func setup(
         with text: NSAttributedString,
         format: RichTextDataFormat
@@ -92,53 +72,25 @@ public protocol RichTextViewComponent: AnyObject,
 
     // MARK: - Functions
 
-    /**
-     Alert a certain title and message.
-
-     - Parameters:
-       - title: The alert title.
-       - message: The alert message.
-       - buttonTitle: The alert button title.
-     */
+    /// Show an alert with a title, message and button text.
     func alert(title: String, message: String, buttonTitle: String)
 
-    /**
-     Copy the current selection.
-     */
+    /// Copy the current selection.
     func copySelection()
 
-    /**
-     Try to redo the latest undone change.
-     */
+    /// Try to redo the latest undone change.
     func redoLatestChange()
 
-    /**
-     Scroll to a certain range.
-
-     - Parameters:
-       - range: The range to scroll to.
-     */
+    /// Scroll to a certain range.
     func scroll(to range: NSRange)
 
-    /**
-     Set the rich text in the text view.
-
-     - Parameters:
-       - text: The rich text to set.
-     */
+    /// Set the rich text in the text view.
     func setRichText(_ text: NSAttributedString)
 
-    /**
-     Set the selected range in the text view.
-
-     - Parameters:
-       - range: The range to set.     
-     */
+    /// Set the selected range in the text view.
     func setSelectedRange(_ range: NSRange)
 
-    /**
-     Undo the latest change.
-     */
+    /// Undo the latest change.
     func undoLatestChange()
 }
 
@@ -147,36 +99,24 @@ public protocol RichTextViewComponent: AnyObject,
 
 public extension RichTextViewComponent {
 
-    /**
-     Alert a title and message, using "OK" as button text.
-
-     - Parameters:
-       - title: The alert title.
-       - message: The alert message.
-     */
+    /// Show an alert with a title, message and OK button.
     func alert(title: String, message: String) {
         alert(title: title, message: message, buttonTitle: "OK")
     }
+    
+    /// Delete all characters in a certain range.
+    func deleteCharacters(in range: NSRange) {
+        mutableAttributedString?.deleteCharacters(in: range)
+    }
 
-    /**
-     Move the text cursor to a certain input index.
-
-     This will use `safeRange(for:)` to cap the index to the
-     available rich text range.
-     */
+    /// Move the text cursor to a certain input index.
     func moveInputCursor(to index: Int) {
         let newRange = NSRange(location: index, length: 0)
         let safeRange = safeRange(for: newRange)
         setSelectedRange(safeRange)
     }
 
-    /**
-     Setup the view with data and a ``RichTextDataFormat``.
-
-     - Parameters:
-       - data: The rich text data to edit with the text view.
-       - format: The rich text format to use.
-     */
+    /// Setup the view with data and a ``RichTextDataFormat``.
     func setup(
         with data: Data,
         format: RichTextDataFormat
@@ -185,13 +125,10 @@ public extension RichTextViewComponent {
         setup(with: string, format: format)
     }
 
-    /**
-     Get the image configuration to use for a certain format.
-
-     - Parameters:
-       - format: The format to get a configuration for.
-     */
-    func standardImageConfiguration(for format: RichTextDataFormat) -> RichTextImageConfiguration {
+    /// Get the image configuration for a certain format.
+    func standardImageConfiguration(
+        for format: RichTextDataFormat
+    ) -> RichTextImageConfiguration {
         let insertConfig = standardImageInsertConfiguration(for: format)
         return RichTextImageConfiguration(
             pasteConfiguration: insertConfig,
@@ -199,42 +136,29 @@ public extension RichTextViewComponent {
             maxImageSize: (width: .frame, height: .frame))
     }
 
-    /**
-     Get the image configuration to use for a certain format.
-
-     - Parameters:
-       - format: The format to get a configuration for.
-     */
-    func standardImageInsertConfiguration(for format: RichTextDataFormat) -> RichTextImageInsertConfiguration {
+    /// Get the image insert config for a certain format.
+    func standardImageInsertConfiguration(
+        for format: RichTextDataFormat
+    ) -> RichTextImageInsertConfiguration {
         format.supportsImages ? .enabled : .disabled
     }
 }
 
 internal extension RichTextViewComponent {
 
-    /**
-     This can be called to setup the initial font size for a
-     text view component.
-     */
+    /// This can be called to setup the initial font size.
     func setupInitialFontSize() {
         let font = FontRepresentable.standardRichTextFont
         let size = font.pointSize
         setCurrentFontSize(to: size)
     }
 
-    /**
-     This can be called to setup an initial text color for a
-     text view component.
-
-     To avoid incorrectly coloring an already formatted text,
-     the action is only called if the provided text is empty.
-     */
+    /// This can be called to setup an initial text color.
     func trySetupInitialTextColor(
         for text: NSAttributedString,
         _ action: () -> Void
     ) {
-        if text.string.isEmpty {
-            action()
-        }
+        guard text.string.isEmpty else { return }
+        action()
     }
 }